fix(python): add typed-error wrapping parity across 8 SDK methods
Move _SCP_CODE_RE and _coded_bridge_error from trust.py to errors.py so
both scp.py and trust.py can share the canonical bridge-error translator.

Wrap identity_execute_recovery, identity_remove, context_member_count,
context_send, ucan_validate, event_log_query, outlet_invoke, and
governance_propose through _coded_bridge_error — the Python analog of the
TypeScript mapBridgeError try/catch added to the same 8 methods in this PR.
Closes the TS↔Python typed-error parity gap flagged in the API-design review.
